Update admin.php
did not work on 6.4.2 WordPress. 

AH01071: Got error 'PHP message: PHP Fatal error:  Uncaught TypeError: unserialize(): Argument #1 ($data) must be of type string, array given in /wp-content/plugins/wp-add-mime-types/includes/admin.php:27

The error message indicates that unserialize() is expecting a string as its argument, but it's receiving an array. To fix this issue, removed unserialize() as if it's already an array, so it will be assigned directly to $mime_type_values